+++ packages/web3-eth/src/utils/try_send_transaction.ts
@@ -17,19 +17,16 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import { Web3Context } from 'web3-core';
 import { EthExecutionAPI, Bytes } from 'web3-types';
 import { AsyncFunction, rejectIfTimeout } from 'web3-utils';
-
 import { TransactionSendTimeoutError } from 'web3-errors';
-import { NUMBER_DATA_FORMAT } from '../constants';
+
 // eslint-disable-next-line import/no-cycle
 import { rejectIfBlockTimeout } from './reject_if_block_timeout';
-// eslint-disable-next-line import/no-cycle
-import { getBlockNumber } from '../rpc_method_wrappers';
 
 /**
  * An internal function to send a transaction or throws if sending did not finish during the timeout during the blocks-timeout.
- * @param web3Context the context to read the configurations from
- * @param sendTransactionFunc the function that will send the transaction (could be sendTransaction or sendRawTransaction)
- * @param transactionHash to be used inside the exception message if there will be any exceptions.
+ * @param web3Context - the context to read the configurations from
+ * @param sendTransactionFunc - the function that will send the transaction (could be sendTransaction or sendRawTransaction)
+ * @param transactionHash - to be used inside the exception message if there will be any exceptions.
  * @returns the Promise<string> returned by the `sendTransactionFunc`.
  */
 export async function trySendTransaction(
@@ -37,8 +34,6 @@ export async function trySendTransaction(
 	sendTransactionFunc: AsyncFunction<string>,
 	transactionHash?: Bytes,
 ): Promise<string> {
-	const pollingInterval = web3Context.transactionPollingInterval;
-
 	const [timeoutId, rejectOnTimeout] = rejectIfTimeout(
 		web3Context.transactionSendTimeout,
 		new TransactionSendTimeoutError({
@@ -47,18 +42,20 @@ export async function trySendTransaction(
 		}),
 	);
 
-	const starterBlockNumber = await getBlockNumber(web3Context, NUMBER_DATA_FORMAT);
-	const [intervalId, rejectOnBlockTimeout] = rejectIfBlockTimeout(
+	const [rejectOnBlockTimeout, blockTimeoutResourceCleaner] = await rejectIfBlockTimeout(
 		web3Context,
-		starterBlockNumber,
-		pollingInterval,
 		transactionHash,
 	);
 
 	try {
-		return await Promise.race([sendTransactionFunc(), rejectOnTimeout, rejectOnBlockTimeout]);
+		// If an error happened here, do not catch it, just clear the resources before raising it to the caller function.
+		return await Promise.race([
+			sendTransactionFunc(), // this is the function that will send the transaction
+			rejectOnTimeout, // this will throw an error on Transaction Send Timeout
+			rejectOnBlockTimeout, // this will throw an error on Transaction Block Timeout
+		]);
 	} finally {
 		clearTimeout(timeoutId);
-		clearInterval(intervalId);
+		blockTimeoutResourceCleaner.clean();
 	}
 }